LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/manual/filter/offscreencanvas.filter.w-expected.txt

CONSOLE MESSAGE: Error: assert_array_equals: The image data generated by filter blur(10px) should be the same for both OffscreenCanvas and regular canvas expected property 0 to be 0 but got 113 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "113,38,28,27,111,34,26,30,108,39,23,33,106,43,28,36,105,46,26,39")

Harness Error (FAIL), message = Error: assert_array_equals: The image data generated by filter blur(10px) should be the same for both OffscreenCanvas and regular canvas expected property 0 to be 0 but got 113 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "113,38,28,27,111,34,26,30,108,39,23,33,106,43,28,36,105,46,26,39")

TIMEOUT offscreencanvas Test timed out

Harness Error (FAIL), message = Error: assert_array_equals: The image data generated by filter blur(10px) should be the same for both OffscreenCanvas and regular canvas expected property 0 to be 0 but got 113 (expected array object "0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0" got object "113,38,28,27,111,34,26,30,108,39,23,33,106,43,28,36,105,46,26,39")

TIMEOUT offscreencanvas Test timed out
